Fix config and window size.
Config: remove settings that adjust the time range when looking up items
in the data source.

Window size: make it 128x128 and centered at the vessel instead of 1024x1024
and aligned on a grid. Also increase the time range so that it can match
successfully, this may be because of using different timestamps reported by
the Sentinel-2 API/image in different places.