Add --no_log and --force_log options.
--no_log can be used for debug runs that you don't want appearing on W&B, e.g. if there
may be errors in the script that happen after the W&B run starts and you want to fix
those first, or if reviewing a PR and testing the training to make sure it runs.

--force_log can be used during `model test` to create W&B run so the metrics can be
viewed there, the primary use case for now is if you log a confusion matrix, if it wasn't
there during training or if val/test are different splits then it needs to be in W&B to
visualize easily.