Remove maldives ecosystem mapping training pipelines.
These aren't used since the README already says to just run `model fit` directly
and I think the train pipeline is kind of pointless since it's just one line and
really should just reference the config file (so even if it isn't documented in
a README, it should be easy to reproduce).

This resolves #14.

This also removes the class from the pipeline definitions so it just needs to be
a function, the class was not used.