Add time series and multimodal configs for marine_infra and solar_farm.
I also updated the launcher code to accept specifying a list of configs. This enables
reducing duplication between some of the config files, although makes it more complicated
to start the experiments since you need to specify a list of configs to get the right
combination (this is documented in the README files within each task dir in data/helios
though).